TASK RUNNERS NPM. Tested the 'npm run static:img:compressed' with tinify png compression.
This option works on Mac, Windows, and Ubuntu, but requires a key and is limited to 500 compressions per month.
You only pay for what you use. The first 500 compressions each month are free. You will only be billed if you compress more than 500 images.
Despite this, this is the best option, since it compiles in all operating systems the first time and without headaches due to incompatibility of npm packages.
https://tinypng.com/developers
https://tinypng.com/dashboard/api